XMLModel per-field load funcs w repeating fields; mods:topic
<mods:topic> has two data points: a URL and a string.  There can
also be multiple topics.  XMLModel/XMLForms has so far been built
to accommodate single values, though there are now hacks for
writing values to several places.

With this change, fields can specify functions to be passed to
XMLModel.__init__() so structured data can be loaded from a parent
tag.

We are starting to resort to fugly hacks here.  I think this
XMLModel/XMLForms model has been taken as far as it can go.