run source before important cmds in debian/package.py
The pip install command is running inside the general buildbot
virtualenv rather than the temporary one that is being built for
the package.  I think this is because the source command is being
run in a separate subcall() function call.  This change prepends
the source command to the pip install command.