keep base path in session instead of config file
The collection base path was specified in a settings file as
DDR_BASE_PATH or LOCAL_BASE_PATH.  This was fine for development
but we don't want to make the partners edit a settings file if
they change USB drives.

This change replaces calls to the semi-hard-coded DDR_BASE_PATH
with storage.base_path() which attempts to get a path from the
cache.  The cache timeout is reset every time the variable is
accessed.