add/rm symlink to ddr dir on USB HDD when mounting/unmounting
We don't know in advance what the name of the USB HDD will be
We don't want to store that in a settings file bc we don't want
to make our users have to edit that file.
Nginx needs to have a stable file to use for /media/.
So, symlink to /media/USBHDDNAME/ddr when mounting.