check for existence of thumbnail before trying to display
Some files are truly huge and some simply can't even be used
to generate thumbs.  By default, sorl attempts to make a thumbnail
every time until it succeeds.
The Entity.add_file() method marks whether it was possible to
generate a thumbnail during ingest and indicates whether or not
to even attempt to display a thumbnail.