manually pass shared folder list into new-file form
For some reason when using the standard Django FilePathField
the list of file choices in the shared folder does not update
when new files are added.  It only updates when the whole
Django process is restarted.
The field seemed to cache values (new values would be visible
in the template but not in the form) even when choices were
supplied by a custom function.
With this change, the list of files in the directory is created
manually and the list is manually pushed in to the form in the
view.