use introspection to find prepare_/process_/ead_/mets_
Many collection and entity fields contain structured data that
requires preparation before it can be used in a form, and post-
processing once it comes back from the form.
Many fields require special functions to insert their data into EAD
or METS XML.
These functions had to be explicitly linked to from COLLECTION_FIELDS
and ENTITY_FIELDS.

With this commit, functions with a field name and a prefix will
automatically be executed if they are present in
ddrlocal.models.collection or ddrlocal.models.entity.
Examples:
- prepare_persons
- process_topics
- ead_created, ead_lastmod
- mets_id