merge entity filemeta into entity file
THIS IS A DESTRUCTIVE CHANGE!!!
BACKWARDS COMPATIBILITY IS NOT SUPPORTED!!!

Entity.files is list of dicts, each with metadata for one file.
Previously, higher-level metadata was kept in entity.filemeta,
which was a dict of dicts keyed to files' SHA1 hashes.
This seemed like a good idea at the time, since it mirrored the
DDR's separation into lower-level code that was concerned only
with files as blobs, and higher-level code that tracked nameing,
subjects, etc, etc.

This worked until someone uploaded two copies of the same file
to an entity.  The SHA1 hash only counts the *contents* of a file,
not the file name.  Both files had the same name, so they pointed
to the same filemeta dict.

This change merges all the data from filemeta into entity.files.
As noted above, compatibility with earlier versions is not
supported.  You've just been uploading QA TEST data, right?