change DDRFile.path to DDRFile.path_rel: be explicit!
There are a lot of different ".*path*" variables and methods
in this code.  Sometimes it's not obvious whether paths are
absolute or relative.
".path" is often used in Django code to mean an absolute path.
It's better to be more explicit.

In the case of DDRFile, we are storing path_rel and *not*
storing path_abs, because we don't know where in the future the
repo will be located and we don't want to cause errors.