mv file metadata to individual FILENAME.json files
IMPORTANT: DESTRUCTIVE AND INCOMPLETE EDIT!

We need to be able to remove certain entity files and their
metadata from the collection repo before publishing.
In order to do so, we have to move the file metadata from the
entity.json to individual FILENAME.json files.

This should hopefully result in cleaner and more consistent
code, after all the dust settles.  However, there is a lot of
dust in the air as of thie commit!