get list of accessible orgs from gitolite instead of settings
The list of repos/orgs a user had access to was set in ddr.cfg.
However, ddr.cfg is shared across all VMs so really doesn't
reflect local config.
For now, solution is to do a "ssh git@SERVER info" and parse
the list that comes back from Gitolite (if any), and make the
repos/orgs list from that.
Thus, the repos/orgs a user has access to depends on the VM's
SSL keys, not their workbench login.
issue #34