first stab at unittesting using Django's test browser simulator
These tests are not an exhaustive unit-testing implementation.
It basically just browses the various pages and confirms that
none of the pages errors out.  Login and logout are implemented
but no editing is done.

Django's test client visits URIs, retrieves HTTP responses and
page text as well as the context returned by the view and some
template information.