lots of work on testing collection/entity edit views
Having test_NN_login and _logout functions in each test class did
not work because we are not using a normal Django login process.

We still have login and logout test functions; these seem to test
login/logout successfully.  Login doesn't carry over to other
test functions, however, so all the extra login/logout functions
were removed, leaving only the first one of each.

In edit test functions, the test client was not logged in.
Perhaps sessions work differently from the way the Django docs say
they do.  Or maybe the webui.views.decorator.login_required
doesn't work the way we expect.
The login_required decorator checks for "git_name" and "git_mail"
in session, so we write these to session inside the edit test
functions and it works.

For forms, it is not enough to insert only the field with the test
string.  All form fields must be filled.