mv task start datetime out of sessions; add in task list view
We want the task start attrib to be a datetime so templates can
display it, but the datetime causes task serialization to crash.
Adding the startd datetime in webui.tasks.session_tasks_list()
lets the datetime still get added to the session, which cannot
be serialized by Redis.
This change removes the datetime from sessions_tasks_list and
adds it in the tasks list view.