change how gitstatus picks new timestamps after updating repos
After updating repo statuses, pick the next earliest time the repo
can be updated.  Previously this time was some specified interval
in the future plus/minus a random value with a specified margin.
This didn't work if the number of repos was greater than the
interval.  The new method picks a time slightly after the highest
timestamp in the queue plus a minute, or a specified minimum.