gitstatus.tmp_dir no longer mkdirs if Store unmounted
If the gitstatus update background process uses a tmp/ dir to
cache its queue and results.  The tmp_dir function would
automatically create the dir if it was absent.  It would do this
even if no Store was mounted (i.e. no /var/www/media/base/ symlink
was present), which was a problem because then no Store *could*
be mounted.
This change only allows creation of the tmp/ dir if a Store
is mounted (i.e. the symlink is present).