import blank content models if real ones are not available
Content model definitions should be in STORE/ddr/repo_models/.
STORE may not be mounted, but various code still refers to the
modules.  This commit adds dummy modules that do well enough
until a STORE can be mounted.  Code will not actually *use* the
modules because if STORE is unmounted the browser will redirect
to the storage mount URLs.