git-annex v5 everywhere; install from backports on wheezy
When git-annex v5 encounters a repo it automatically upgrades it,
rendering the repo unreadable to earlier versions.
Debian 7 (Wheezy) contains an older version of git-annex than
Debian 8 (Jessie).
Because of our unique setup in the HQ, if a Deb8 VM touches a repo
it gets upgraded and it becomes useless to our Deb7 VMs.

This change will upgrade git-annex on Deb7 machines to v5.

see 8aca1d62b3 on develop.