entity edit view now groks both new,old topics,facility data fmt
In the JSON topics,facility are now represented as list-of-dicts
with separate term and ID subfields.
We're still using Tagmanager for the UI though, which uses the
old format.
So we need to convert new topics,facility data into the old format
for the form and back into the new format afterwards