don't raise exception if module definition commit not in log
When documents are saved, the metadata includes the commit hash of
the model definitions module.  When we load it we indicate whether
the current model definitions are newer/older than the ones used
in the document.
One edge case is when the model definitions are on an unmerged
branch when a document is committed.  In that case we want to
report that fact rather than crashing.

see ddr-cmdln commit f7f8be7