Makefile: remove dependency on apache2 in make-deb
Cgit depends on some sort of web server.  Apparently listing
Nginx later in the dependency list makes dpkg think it should
install apache2, and then it and nginx fight.  So moved nginx
to the beginning of the list.
Apparently just saying "nginx" isn't specific enough, so specified
"nginx-light" as a dependency.