models.OBJECT.save: Don't mix Django with ddr-cmdln behavior
Previous versions of ddr-cmdln took a Django-style cleaned_data arg
and ran form_post on the data.  This was harder to test that it
should be, and mixed behavior that should have stayed in this
Django app.
This change keeps the Django forms cleaned_data stuff here and
uses the updated model.OBJECT.save from ddr-cmdln.