Makefile: Move apt-install calls to install-dependencies task
apt-install is called in the install-ddr-cmdln and install-ddr-local
tasks.  These were moved to a single install-dependencies task
that is triggered in install-app

This is a slight optimization but the real reason was that the
apt-install call in install-ddr-local was clobbering the
ImageMagick policy.xml file installed in install-ddr-cmdln and
preventing build.sr.ht from working.