Makefile: Add git safe.directory to system-wide config
Git now requires the safe.directory setting when accessing
repositories that are not owned by the current user.
This is a problem for our apps that want to run git commands
to report their current version.
Makefile is able to run these commands but Debian packages can't.
Also for some reason the ddr user does not seem to respond to
this setting when running Celery under Supervisord.
So, this blunt force solution: write configs to /etc/gitconfig.
Change is applied both during make-install and package install.